fix(JWTAuthenticationProvider.java): fix getting claims from JWT token and setting user properties correctly
The code was updated to fix the issue with getting claims from the JWT token. Previously, the code was using the wrong variable name `jwtAuthentication.getToken()` instead of `token` to get the claims. This has been fixed by using the correct variable name.

Additionally, the code was updated to set the user properties correctly. The email property was being set using the wrong key `claims.get("email")` instead of `claims.get("sub")`. This has been fixed to set the email property correctly.

Furthermore, the code was updated to set the userId property correctly. The userId was being retrieved as an Integer from the claims and then casted to Long, which could cause a ClassCastException. This has been fixed by retrieving the userId as an Integer and then converting it to Long using `Long.valueOf()` method.